requirements.txt
flask==2.3.3
gspread==5.12.0
google-auth==2.23.0
pandas==2.1.0
selenium==4.12.0
webdriver-manager==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0